Simplify Dashboard and Implement Splash Screen API
