+++ app/build.gradle
@@ -8,7 +8,7 @@ plugins {
 def keystorePropertiesFile = rootProject.file("keystore.properties")
 
 android {
-    compileSdk 31
+    compileSdk 32
     signingConfigs {
         release {
             if (keystorePropertiesFile.exists()) {
@@ -24,7 +24,7 @@ android {
     defaultConfig {
         applicationId 'org.metabrainz.android'
         minSdk 21
-        targetSdk 31
+        targetSdk 32
         versionCode 46
         versionName "5.0.5"
 
@@ -86,6 +86,7 @@ dependencies {
     implementation 'androidx.recyclerview:recyclerview:1.2.1'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
     implementation 'androidx.preference:preference-ktx:1.2.0'
+    implementation 'androidx.core:core-splashscreen:1.0.0-beta02'
 
     //Web Service Setup
     implementation 'com.google.code.gson:gson:2.9.0'
